svn:2225 RIPL-3 resonances files replace RIPL-2 files previously in EMPIRE. Those files were also used for EGSM LD parameterization. This modification affects the reading of D0 and Gg used in the radiative width renormalization.
The renormalization energy changed to < 2 keV.
warning meesages introduced if Bn < Ucrt - DELTA as the calculation of D0 is not accurate in this case.

Consistency of calculated D0 verified vs those values obtained in the LD fit, as well as experimental D0 values for those nuclei where Ucrt - DELTA < Bn.